+++ extension.js
@@ -3,17 +3,14 @@ const Categories = require('./lib/categories');
 
 function activate(context) {
   const categories = new Categories();
-  const configuredCategories = workspace.getConfiguration("rails-project-navigator.categories");
-
-  configuredCategories.forEach(function(category) {
-    categories.add(category.name, category.path, category.exclusionGlob);
-  });
 
   var disposable = commands.registerCommand('railsProjectNavigator.navigate', function () {
-    window.showQuickPick(categories.names(), {placeholder: 'Select a Category'}).then((selected) => {
-      if (!selected) { return; }
-
-      categories.showFilesFor(selected);
+    return new Promise((res, rej) => {
+      categories.showCategories()
+        .then(categories.getFilesFor.bind(categories), rej)
+        .then(categories.showFiles.bind(categories), rej)
+        .then(categories.openFile.bind(categories), rej)
+        .then(res, rej);
     });
   });